npHard/requirements.txt
absl-py==0.8.1
astor==0.8.0
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.8.1
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.8
cycler==0.10.0
dash==0.35.1
dash-bootstrap-components==0.3.2
dash-core-components==0.41.0
dash-dangerously-set-inner-html==0.0.2
dash-daq==0.1.4
dash-html-components==0.13.4
dash-renderer==0.13.0
dash-table==3.1.11
dash-table-experiments==0.6.0
Flask==1.1.1
Flask-Caching==1.7.2
Flask-Compress==1.4.0
future==0.18.1
gast==0.3.2
grasia-dash-components==0.3.4
grpcio==1.24.1
gunicorn==19.9.0
h5py==2.10.0
hachoir==3.0a5
idna==2.8
imageio==2.6.1
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mock==3.0.5
more-itertools==7.2.0
numpy==1.17.3
opencv-python==4.1.1.26
packaging==19.2
percy==2.0.2
Pillow==6.2.0
plotly==4.2.1
pluggy==0.13.0
protobuf==3.10.0
py==1.8.0
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.4.2
pytest==5.2.1
pytest-mock==1.11.1
pytest-sugar==0.9.2
python-dateutil==2.8.0
pyttsx3==2.71
PyYAML==5.1.2
requests==2.22.0
retrying==1.3.3
scipy==1.3.1
selenium==3.141.0
six==1.12.0
soupsieve==1.9.4
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.25.6
waitress==1.3.1
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
